@@ -13,33 +13,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.14",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/forms": "8.2.14",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
+    "@angular/animations": "9.0.1",
+    "@angular/common": "9.0.1",
+    "@angular/compiler": "9.0.1",
+    "@angular/core": "9.0.1",
+    "@angular/forms": "9.0.1",
+    "@angular/platform-browser": "9.0.1",
+    "@angular/platform-browser-dynamic": "9.0.1",
+    "@angular/router": "9.0.1",
     "bootstrap": "4.1.x",
     "core-js": "2.5.x",
     "intl": "1.2.x",
     "moment": "^2.24.0",
     "rxjs": "6.5.4",
     "rxjs-compat": "^6.0.0-rc.0",
-    "tslib": "1.9.x",
-    "zone.js": "~0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.25",
-    "@angular/cli": "8.3.25",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/language-service": "8.2.14",
+    "@angular-devkit/build-angular": "~0.900.2",
+    "@angular/cli": "9.0.2",
+    "@angular/compiler-cli": "9.0.1",
+    "@angular/language-service": "9.0.1",
     "@types/jasmine": "2.8.x",
     "@types/jasminewd2": "2.0.x",
     "@types/moment": "^2.13.0",
-    "@types/node": "9.6.x",
-    "codelyzer": "^5.0.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "3.1.x",
     "jasmine-spec-reporter": "4.2.x",
     "karma": "2.0.x",
@@ -53,6 +53,6 @@
     "ts-mockito": "2.3.x",
     "ts-node": "5.0.x",
     "tslint": "5.9.x",
-    "typescript": "3.5.3"
+    "typescript": "3.7.5"
   }
 }
